LayoutTests/fast/hidpi/image-srcset-invalid-inputs-correct-src-expected.txt

PASS document.getElementById("foo").clientWidth==400 is true
This test passes if this img tag below is a green square regardless of the scale factor. It ensures that invalid inputs from srcset are ignored, and src is selected (since it's the only candidate left)
